fix(db): keep avatars and profile working when Postgres is down
Avatar reads (local photo lookup, trombinoscope batch, colour variant)
now fall back instead of throwing: a Postgres outage shows generated
initials in the default colour rather than failing /profile, the
trombinoscope and the public /avatars endpoint used by Authentik and
BookStack. The profile's audit history degrades to empty the same way.

Also switches wishlist_items.estimated_amount from REAL (4-byte float in
Postgres) to DOUBLE PRECISION, matching SQLite's 8-byte REAL.

+++ src/lib/server/migrations.ts
@@ -57,7 +57,9 @@ const migrations: Migration[] = [
 					description TEXT,
 					link TEXT,
 					quantity INTEGER NOT NULL DEFAULT 1,
-					estimated_amount REAL,
+					-- DOUBLE PRECISION, not REAL: Postgres' REAL is 4 bytes (~7 significant digits),
+					-- unlike SQLite's 8-byte REAL this column was first written for.
+					estimated_amount DOUBLE PRECISION,
 					type TEXT NOT NULL
 				)
 			`;
